Create NotificationDelegate for notification handling on iOS
